+++ src/mrklkit/lruntime.h
@@ -166,7 +166,7 @@ void **mrklkit_rt_get_struct_item_addr(rt_struct_t *, int64_t);
 #define MRKLKIT_RT_GET_STRUCT_ITEM_ADDR(val, idx) ((val)->fields + (idx))
 int64_t mrklkit_rt_get_struct_item_int(rt_struct_t *, int64_t);
 double mrklkit_rt_get_struct_item_float(rt_struct_t *, int64_t);
-int64_t mrklkit_rt_get_struct_item_bool(rt_struct_t *, int64_t);
+int8_t mrklkit_rt_get_struct_item_bool(rt_struct_t *, int64_t);
 bytes_t *mrklkit_rt_get_struct_item_str(rt_struct_t *, int64_t);
 rt_array_t *mrklkit_rt_get_struct_item_array(rt_struct_t *, int64_t);
 rt_dict_t *mrklkit_rt_get_struct_item_dict(rt_struct_t *, int64_t);
